+++ src/doorman/Byte.hpp
@@ -1,0 +1,3 @@
+#pragma once
+
+using byte = unsigned char;
